+++ spec/monolens/object/test_select.rb
@@ -75,6 +75,91 @@ describe Monolens, 'object.select' do
     end
   end
 
+  context 'when using strategy: first' do
+    subject do
+      Monolens.lens('object.select' => {
+        defn: {
+          name: [:firstname, :lastname],
+          status: :priority
+        },
+        strategy: 'first',
+        on_missing: on_missing
+      }.compact)
+    end
+
+    context 'without on_missing' do
+      let(:on_missing) do
+        nil
+      end
+
+      it 'works as expected when first option is present' do
+        input = {
+          firstname: 'Bernard',
+          priority: 12
+        }
+        expected = {
+          name: 'Bernard',
+          status: 12
+        }
+        expect(subject.call(input)).to eql(expected)
+      end
+
+      it 'works as expected when second is present' do
+        input = {
+          lastname: 'Lambeau',
+          priority: 12
+        }
+        expected = {
+          name: 'Lambeau',
+          status: 12
+        }
+        expect(subject.call(input)).to eql(expected)
+      end
+
+      it 'fails when none is present' do
+        input = {
+          priority: 12
+        }
+        expect {
+          subject.call(input)
+        }.to raise_error(Monolens::Error)
+      end
+    end
+
+    context 'with on_missing: skip' do
+      let(:on_missing) do
+        :skip
+      end
+
+      it 'works as expected when missing' do
+        input = {
+          priority: 12
+        }
+        expected = {
+          status: 12
+        }
+        expect(subject.call(input)).to eql(expected)
+      end
+    end
+
+    context 'with on_missing: null' do
+      let(:on_missing) do
+        :null
+      end
+
+      it 'works as expected when missing' do
+        input = {
+          priority: 12
+        }
+        expected = {
+          name: nil,
+          status: 12
+        }
+        expect(subject.call(input)).to eql(expected)
+      end
+    end
+  end
+
   context 'when using an array as selection' do
     subject do
       Monolens.lens('object.select' => {
@@ -186,7 +271,6 @@ describe Monolens, 'object.select' do
 
     subject do
       lens.call(input)
-      nil
     rescue Monolens::LensError => ex
       ex
     end
@@ -196,7 +280,7 @@ describe Monolens, 'object.select' do
     end
 
     it 'correctly updates the location' do
-      expect(subject.location).to eql([:status])
+      expect(subject.location).to eql(['status'])
     end
   end
 end